post-process: refactor calltree out of html
Ref: https://github.com/ossf/fuzz-introspector/issues/229
